+++ Lib/Composition/CompilationResultMessage.cs
@@ -4,6 +4,7 @@
     {
         public string FileName;
         public bool IsError;
+        public int Code;
         public string Text;
         /// startLine, startCol, endLine, endCol all one based
         public int[] Pos;